Split HTTPDatabase into focused files
Move each protocol conformance (RecordWriter, RecordReader, RecordLocator, ActivityReader, MetricSeriesReader) out of the monolithic HTTPDatabase.swift into its own Hosted<Protocol>.swift file, leaving HTTPDatabase.swift with only the transport core (the struct, request plumbing, and HTTPDatabaseError).
